+++ lib/src/ui/inspector_overlay.dart
@@ -68,7 +68,9 @@ class _HttpWatcherOverlayState extends State<HttpWatcherOverlay> {
     super.dispose();
   }
 
-  void _refresh() => setState(() {});
+  void _refresh() {
+    if (mounted) setState(() {});
+  }
 
   @override
   Widget build(BuildContext context) {